Camera obstruction fading
Added box collider and kinematic rigidbody to camera.
Implemented some shader code that allows changing the standard shader's rendering mode during runtime.
Works in the editor; HOWEVER, there must be at least one active object in any of the build scenes using a material with the desired shader settings in order to reflect the rendering mode change during actual running of the build. For example, the control_test scene has a primitive, talking box that is using the Fade rendering mode. This allows the Fade rendering mode on any other object to be accessed through scripts when running a real build.
Currently, objects that we wish to fade must be tagged with "Environment."