Menus pull request into master (#4)
* Created UMenu Utility Class

UMenu provides classes and functions common to different menus.

- It provides a MenuOption struct to use for defining each option in a menu.

- It provides a MenuMovement class for packaging and setting flags for menu traversal.

* Pausing now works correctly

Fixed bug causing game to immediately unpause after pausing.

- Pause flag was still set after entering the pause menu, which exits the pause menu when the pause flag is set.

- To fix, all button flags are unset before pausing the game.

* Added non-functional Pause Menu

Added non-functioning pause menu and the resources associated with it. The menu can be traversed with the the horizontal and vertical axes, but the options cannot be selected.

- Added a new Resources folder which holds the images needed for the pause menu. (The Resources folder can be accessed in the code with Resources.Load())

- Added an image prefab to Niko's prefab so that the menu system can instantiate Images.

* Finishing previous commit

Previous commit did not take the right files.

* Options can be selected

Options can now be selected, although no actions have been written for the buttons.

- Added UMenu.clearOptions(List<UMenu.MenuOptions>) function which destroys the images of the passed in menu and clears the rest of the option data.

* Added Resume Action

Selecting Resume in the main menu now resumes the game!

* Camera Changes

Camera is much smoother now, and is ready to be inverted as soon as player settings are implemented.

- Camera now runs from FixedUpdate instead of Update. It was really choppy running in Update.

- Mouse input is no longer inverted. Left joystick is still inverted, because the mouse and joystick orientation are opposite.

- ControlManager now tracks the inversion status of X and Y axes.

* Forgot to save Unity scene

Adding Unity scene to commit.

* Added empty functions for menu

Menu now has empty function declarations for Save, Load, Options, and Title Screen choices.

- Options and Title Screen should be implemented before the end of the semester.

* Added Options Menu

Added Options menu, which is accessible from the pause menu.

- Can go back and forth between pause menu and options menu.

- Currently does nothing.

* Implemented Camera Inversion Settings

It is now possible to invert the camera through the Options menu.

- The menu code needs a refactor if any more submenus are to be added, but it isn't worth the time for now.

* Added values to UGen Enums

Added some values to the layer enumerators in UGen.

* Added comment for some new camera code

Implemented alpha reduction of objects that are between the camera and the player.

- Promptly commented the code out, because it's slow and it looks terrible.

- Left it in commented to keep the idea around.

* Added comment

Added a comment in OrbitCamera

* Camera obstruction fading

Added box collider and kinematic rigidbody to camera.
Implemented some shader code that allows changing the standard shader's rendering mode during runtime.
Works in the editor; HOWEVER, there must be at least one active object in any of the build scenes using a material with the desired shader settings in order to reflect the rendering mode change during actual running of the build. For example, the control_test scene has a primitive, talking box that is using the Fade rendering mode. This allows the Fade rendering mode on any other object to be accessed through scripts when running a real build.
Currently, objects that we wish to fade must be tagged with "Environment."